pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurosynth"
version = "0.1.0"
description = "NeuroSynth data lakehouse and ingestion pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic==2.7.4",
  "pydantic-settings==2.3.4",
  "structlog==24.4.0",
  "tenacity==9.0.0",
  "aiofiles==24.1.0",
  "httpx==0.27.0",
  "asyncpg==0.29.0",
  "psycopg[binary]==3.2.1",
  "boto3==1.34.152",
  "confluent-kafka==2.5.0",
  "pyiceberg[pyarrow,pandas]==0.7.1",
  "pyarrow==16.1.0",
  "pandas==2.2.2",
  "numpy==1.26.4",
  "scipy==1.13.1",
  "neo4j==5.22.0",
  "pydicom==2.4.4",
  "SimpleITK==2.3.1",
  "asyncio-mqtt==0.16.2",
  "nibabel==5.2.1",
  "nilearn==0.10.4",
  "scikit-learn==1.4.2",
  "torch==2.3.1",
  "torch-geometric==2.5.3",
  "torch-scatter==2.1.2",
  "torch-sparse==0.6.18",
  "torchdiffeq==0.2.4",
  "networkx==3.3",
  "mlflow==2.14.3",
  "lifelines==0.29.0",
  "plotly==5.22.0",
  "transformers==4.40.2",
  "scikit-allel==1.3.8",
  "biopython==1.84",
  "pysam==0.22.1",
  "h5py==3.11.0",
  "hdf5plugin==4.4.0",
  "pyliftover==0.4.1",
  "opacus==1.5.2",
  "hail==0.2.131",
  "pytorch-forecasting==1.1.1",
  "pytorch-lightning==2.2.5",
  "optuna==3.6.1",
  "shap==0.45.1",
  "pandera==0.20.2",
  "dowhy==0.11",
  "econml==0.14.1",
  "causal-learn==0.1.3.8",
  "pgmpy==0.1.26",
  "matplotlib==3.8.4",
  "statsmodels==0.14.2",
  "dask==2024.5.2",
  "peft==0.10.0",
  "trl==0.8.6",
  "vllm==0.4.3",
  "outlines==0.0.40",
  "langchain==0.2.9",
  "llama-index==0.10.56",
  "qdrant-client==1.9.2",
  "bitsandbytes==0.43.1",
  "biopython==1.84",
  "datasets==2.20.0",
  "sentence-transformers==3.0.1",
  "rank-bm25==0.2.2",
  "rouge-score==0.1.2",
  "bert-score==0.3.13",
  "presidio-analyzer==2.2.355",
  "fastapi==0.111.1",
  "uvicorn==0.30.1",
  "prometheus-client==0.20.0",
  "prometheus-fastapi-instrumentator==7.0.0",
  "slowapi==0.1.9",
  "pyjwt==2.8.0",
  "python-jose==3.3.0",
  "celery==5.4.0",
  "redis==5.0.7",
  "kfp==2.8.0",
  "evidently==0.4.37",
  "locust==2.30.1",
]

[project.optional-dependencies]
test = [
  "pytest==8.3.2",
  "pytest-asyncio==0.23.8",
  "respx==0.21.1",
  "fakeredis==2.25.1",
  "pytest-cov==5.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q --cov=neurosynth --cov-report=term-missing"
asyncio_mode = "auto"
